Refactored Navigation extensively
