+++ src/components/Navigation/Bar.js
@@ -1,4 +1,4 @@
-import React, { useCallback } from "react";
+import React from "react";
 import styled from "styled-components";
 import { getThemeProp } from "../../utils";
 import Tab from "./Tab";
@@ -17,31 +17,27 @@ export const TabBar = styled.div`
 	text-transform: uppercase;
 `;
 
-const Bar = ({ pages, close, moduleName, moduleHref }) => {
-	const innerClose = useCallback(close(moduleName, moduleHref), [
-		close,
-		moduleName,
-		moduleHref,
-	]);
-	return (
-		<TabBar>
-			{pages.map(
-				({ href, mappedFrom, label, active, icon, outsideScope }, index) => (
-					<Tab
-						key={href}
-						module={index === 0}
-						icon={icon}
-						href={href}
-						mappedFrom={mappedFrom}
-						label={label}
-						active={active}
-						close={innerClose}
-						outsideScope={outsideScope}
-					/>
-				),
-			)}
-		</TabBar>
-	);
-};
+const Bar = ({ pages }) => (
+	<TabBar>
+		{pages.map(
+			(
+				{ href, mappedFrom, label, active, icon, outsideScope, close },
+				index,
+			) => (
+				<Tab
+					key={href}
+					module={index === 0}
+					icon={icon}
+					href={href}
+					mappedFrom={mappedFrom}
+					label={label}
+					active={active}
+					close={close}
+					outsideScope={outsideScope}
+				/>
+			),
+		)}
+	</TabBar>
+);
 
 export default Bar;
