+++ src/components/Navigation/Bar.test.js
@@ -2,15 +2,13 @@ import React from "react";
 import { Provider } from "react-redux";
 import { MemoryRouter } from "react-router-dom";
 import sinon from "sinon";
-import { getClassSelector } from "../../utils/testUtils";
-import Tab, { CloseIcon } from "./Tab";
+import Tab from "./Tab";
 import Bar, { TabBar } from "./Bar";
 
 describe("Bar", () => {
-	let close, innerClose;
+	let closers;
 	beforeEach(() => {
-		innerClose = sinon.spy(() => () => {}).named("innerClose");
-		close = sinon.spy(() => innerClose).named("close");
+		closers = [0, 1, 2, 3].map(id => sinon.spy().named("close" + id));
 	});
 
 	it("renders a bar containing tabs", () =>
@@ -35,25 +33,28 @@ describe("Bar", () => {
 								href: "/Foo/modu/1",
 								mappedFrom: "/Foo/modu/1",
 								label: "Page 1",
+								close: closers[0],
 							},
 							{
 								href: "/Foo/modu/2",
 								mappedFrom: "/Foo/modu/2",
 								label: "Page 2",
+								close: closers[1],
 								active: true,
 							},
 							{
 								href: "/Foo/modu/3",
 								mappedFrom: "/Foo/modu/3",
 								label: "Page 3",
+								close: closers[2],
 							},
 							{
 								href: "/Foo/modu/4",
 								mappedFrom: "/Foo/modu/4",
 								label: "Page 4",
+								close: closers[3],
 							},
 						]}
-						close={close}
 					/>
 				</MemoryRouter>
 			</Provider>,
@@ -81,14 +82,14 @@ describe("Bar", () => {
 							href="/Foo/modu/1"
 							mappedFrom="/Foo/modu/1"
 							label="Page 1"
-							close={innerClose}
+							close={closers[0]}
 						/>
 						<Tab
 							key="/Foo/modu/2"
 							href="/Foo/modu/2"
 							mappedFrom="/Foo/modu/2"
 							label="Page 2"
-							close={innerClose}
+							close={closers[1]}
 							active
 						/>
 						<Tab
@@ -96,61 +97,17 @@ describe("Bar", () => {
 							href="/Foo/modu/3"
 							mappedFrom="/Foo/modu/3"
 							label="Page 3"
-							close={innerClose}
+							close={closers[2]}
 						/>
 						<Tab
 							key="/Foo/modu/4"
 							href="/Foo/modu/4"
 							mappedFrom="/Foo/modu/4"
 							label="Page 4"
-							close={innerClose}
+							close={closers[3]}
 						/>
 					</TabBar>
 				</MemoryRouter>
 			</Provider>,
 		));
-
-	it("curries the close handler with module name and href", () =>
-		expect(
-			<Provider
-				store={{
-					subscribe: () => {},
-					dispatch: () => {},
-					getState: () => ({}),
-				}}
-			>
-				<MemoryRouter>
-					<Bar
-						pages={[
-							{
-								icon: "test",
-								label: "A module",
-								href: "/Foo/modu",
-								mappedFrom: "/Foo/modu",
-							},
-							{
-								href: "/Foo/modu/1",
-								mappedFrom: "/Foo/modu/1",
-								label: "Page 1",
-							},
-						]}
-						close={close}
-						moduleName="modu"
-						moduleHref="/Foo/modu"
-					/>
-				</MemoryRouter>
-			</Provider>,
-			"when mounted",
-			"with event",
-			{
-				type: "click",
-				target: getClassSelector(<CloseIcon />),
-			},
-		)
-			.then(() =>
-				expect(close, "to have calls satisfying", [
-					{ args: ["modu", "/Foo/modu"] },
-				]),
-			)
-			.then(() => expect(innerClose, "was called")));
 });
